Jan 1, 2025. Happy New Year!
Commit with a bug found. Bug is not programmatic, I guess, the UN librarians have changed the data on the website, because the November resolution has been grabbed only today and it failed with the error:

ERROR happened while populating database: insert or update on table "resolution_agenda" violates foreign key constraint "resolution_agenda_resolution_id_fkey"
DETAIL:  Key (resolution_id)=(4070358) is not present in table "resolution".

The resolution's title is "Security Council resolution 2760 (2024) [on extension of the mandate of the UN Interim Security Force for Abyei (UNISFA) until 15 Nov. 2025]", and this title is in the database already as logs point out, but the title's id (5790) leads to another resolution id - 4066424 which is "Security Council resolution 2761 (2024) [on humanitarian exemptions to asset freeze measures imposed by ISIL (Da'esh) and Al-Qaida sanctions regime]" of 2024-12-06, which was grabbed on Dec 1's commit, this is impossible unless someone has changed data on the website. Solution is to recollect data from the November's backup. Next commit with the corrected data will be published soon.